CDAP-15860 Facebook Ads (SaaS source) - prepare for streaming source.
