+++ src/main/java/io/cdap/plugin/facebook/source/common/config/BaseSourceConfig.java
@@ -17,12 +17,21 @@
 package io.cdap.plugin.facebook.source.common.config;
 
 import com.google.common.base.Strings;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 import io.cdap.cdap.api.annotation.Description;
 import io.cdap.cdap.api.annotation.Macro;
 import io.cdap.cdap.api.annotation.Name;
+import io.cdap.cdap.api.data.schema.Schema;
 import io.cdap.cdap.etl.api.FailureCollector;
 import io.cdap.plugin.common.ReferencePluginConfig;
+import io.cdap.plugin.facebook.source.common.SchemaHelper;
+import io.cdap.plugin.facebook.source.common.exceptions.IllegalInsightsFieldException;
 
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 
 /**
@@ -35,6 +44,9 @@ public class BaseSourceConfig extends ReferencePluginConfig {
   public static final String PROPERTY_AD_SET_ID = "adSetId";
   public static final String PROPERTY_CAMPAIGN_ID = "campaignId";
   public static final String PROPERTY_ACCOUNT_ID = "accountId";
+  public static final String PROPERTY_FIELDS = "fields";
+  public static final String PROPERTY_LEVEL = "level";
+  public static final String PROPERTY_FILTERING = "filtering";
 
   @Name(PROPERTY_ACCESS_TOKEN)
   @Description("Access Token.")
@@ -70,6 +82,29 @@ public class BaseSourceConfig extends ReferencePluginConfig {
   @Nullable
   protected String accountId;
 
+  @Name(PROPERTY_FIELDS)
+  @Description("Fields to get.")
+  @Macro
+  protected String fields;
+
+  @Name(PROPERTY_LEVEL)
+  @Description("Level of request.")
+  @Macro
+  protected String level;
+
+  @Name(PROPERTY_FILTERING)
+  @Description("Filters to query with.")
+  @Nullable
+  @Macro
+  protected String filtering;
+
+  /*
+  Most likely unique delimiter that helps avoid problems with unescaped symbols in complex filters
+  */
+  public static final String FILTERING_DELIMITER = "%!delim@%";
+  private static final Gson gson = new GsonBuilder().create();
+
+  private transient Schema schema = null;
 
   public BaseSourceConfig(String referenceName) {
     super(referenceName);
@@ -98,6 +133,48 @@ public class BaseSourceConfig extends ReferencePluginConfig {
     }
   }
 
+  public List<String> getFields() {
+    if (!Strings.isNullOrEmpty(fields)) {
+      return Arrays.asList(fields.split(","));
+    } else {
+      return Collections.emptyList();
+    }
+  }
+
+  public Schema getSchema() {
+    if (schema == null) {
+      schema = SchemaHelper.buildSchema(getFields(), getBreakdown());
+    }
+    return schema;
+  }
+
+  public String getLevel() {
+    return level;
+  }
+
+  public Breakdowns getBreakdown() {
+    return null;
+  }
+
+  @Nullable
+  public String getFiltering() {
+    if (!Strings.isNullOrEmpty(filtering)) {
+      return gson.toJson(getFilters());
+    } else {
+      return null;
+    }
+  }
+
+  public List<Filter> getFilters() {
+    if (!Strings.isNullOrEmpty(filtering)) {
+      return Arrays.stream(filtering.split(FILTERING_DELIMITER))
+        .map(SourceConfigHelper::parseFilteringItem)
+        .collect(Collectors.toList());
+    } else {
+      return null;
+    }
+  }
+
   public void validate(FailureCollector failureCollector) {
     if (!containsMacro(PROPERTY_ACCESS_TOKEN) && Strings.isNullOrEmpty(accessToken)) {
       failureCollector
@@ -118,6 +195,8 @@ public class BaseSourceConfig extends ReferencePluginConfig {
     }
 
     validateObjectId(failureCollector);
+    validateFields(failureCollector);
+    validateFiltering(failureCollector);
   }
 
   void validateObjectId(FailureCollector failureCollector) {
@@ -151,6 +230,39 @@ public class BaseSourceConfig extends ReferencePluginConfig {
               .withConfigProperty(PROPERTY_ACCOUNT_ID);
           }
           break;
+      }
+    }
+  }
+
+  void validateFields(FailureCollector failureCollector) {
+    if (!containsMacro(PROPERTY_FIELDS)) {
+      if (Strings.isNullOrEmpty(fields) && getFields().size() == 0) {
+        failureCollector
+          .addFailure("At least one field must be specified.", "Specify valid fields.")
+          .withConfigProperty(PROPERTY_FIELDS);
+      } else {
+        getFields().forEach(field -> {
+          try {
+            SchemaHelper.fromName(field);
+          } catch (IllegalInsightsFieldException ex) {
+            failureCollector
+              .addFailure("Invalid field:" + ex.getFieldName(), "Remove invalid field.")
+              .withConfigElement(PROPERTY_FIELDS, ex.getFieldName());
+          }
+        });
+      }
+    }
+  }
+
+  void validateFiltering(FailureCollector failureCollector) {
+    if (!containsMacro(PROPERTY_FILTERING)) {
+      try {
+        getFiltering();
+      } catch (Exception ex) {
+        // all kind of parsing exceptions
+        failureCollector.addFailure("Failed to parse filtering:" + ex.getMessage(), null)
+          .withStacktrace(ex.getStackTrace())
+          .withConfigProperty(PROPERTY_FILTERING);
       }
     }
   }
